requirements.txt
asgiref==3.8.1
Django==5.1.7
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
django-filter==23.5
django-tables2==2.7.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-widget-tweaks==1.5.0
django-allauth==0.61.1
django-debug-toolbar==4.3.0
